+++ cortex-m-rt/tests/compile-fail/pre-init-bad-signature-2.rs
@@ -6,7 +6,7 @@ extern crate panic_halt;
 
 use cortex_m_rt::{entry, pre_init};
 
-#[pre_init]
+#[pre_init] //~ WARNING Use core::arch::global_asm! to define the __pre_init function instead
 unsafe fn foo(undef: i32) {}
 //~^ ERROR `#[pre_init]` function must have signature `unsafe fn()`