+++ cortex-m-rt/tests/compile-fail/pre-init-twice.rs
@@ -6,10 +6,11 @@ extern crate panic_halt;
 
 use cortex_m_rt::{entry, pre_init};
 
-#[pre_init]
+#[pre_init] //~ WARNING Use core::arch::global_asm! to define the __pre_init function instead
 unsafe fn foo() {}
 
 #[pre_init] //~ ERROR symbol `__pre_init` is already defined
+            //~^ WARNING Use core::arch::global_asm! to define the __pre_init function instead
 unsafe fn bar() {}
 
 #[entry]